Split item handling from server code
